fix security parsing and logic. Permit declaration of base 'security'
