+++ openapi-codegen/core/src/main/scala/sttp/tapir/codegen/util/ErrUtils.scala
@@ -1,0 +1,9 @@
+package sttp.tapir.codegen.util
+
+case class Location(path: String, method: String) {
+  override def toString: String = s"${method.toUpperCase} ${path}"
+}
+
+object ErrUtils {
+  def bail(msg: String)(implicit location: Location): Nothing = throw new NotImplementedError(s"$msg at $location")
+}
